feat: refactor listApps + new implementation
